Move NotifyPropertyChanged and PropertyChangedListener into Doctrine\Persistence
