+++ UPGRADE.md
@@ -1,5 +1,11 @@
 # Upgrade to 2.0
 
+## `Doctrine\Common\NotifyPropertyChanged` and `Doctrine\Common\PropertyChangedListener` moved into `Doctrine\Persistence`
+
+The following classes were moved into different namespace:
+ * `Doctrine\Common\NotifyPropertyChanged` -> `Doctrine\Persistence\NotifyPropertyChanged`
+ * `Doctrine\Common\PropertyChangedListener` -> `Doctrine\Persistence\PropertyChangedListener`
+
 ## Namespace renamed to `Doctrine\Persistence`
 
 The namespace has been renamed from `Doctrine\Common\Persistence` to `Doctrine\Persistence`.
